# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Lov": {
        "properties": {
          "apex_item_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apex Item Height"
          },
          "apex_item_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apex Item Width"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_extra": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": true,
            "title": "Display Extra"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": true,
            "title": "Is Active"
          },
          "is_enterable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Is Enterable"
          },
          "is_static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Is Static"
          },
          "is_translatable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Is Translatable"
          },
          "lov_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lov Definition"
          },
          "lov_name": {
            "title": "Lov Name",
            "type": "string"
          },
          "null_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Text"
          },
          "null_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Value"
          },
          "show_null_value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Show Null Value"
          },
          "static_values": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static Values"
          },
          "translation_applicable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Translation Applicable"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "lov_name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "Lov",
        "type": "object"
      },
      "LovCreate": {
        "properties": {
          "apex_item_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apex Item Height"
          },
          "apex_item_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apex Item Width"
          },
          "display_extra": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": true,
            "title": "Display Extra"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": true,
            "title": "Is Active"
          },
          "is_enterable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Is Enterable"
          },
          "is_static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Is Static"
          },
          "is_translatable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Is Translatable"
          },
          "lov_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lov Definition"
          },
          "lov_name": {
            "title": "Lov Name",
            "type": "string"
          },
          "null_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Text"
          },
          "null_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Value"
          },
          "show_null_value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Show Null Value"
          },
          "static_values": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static Values"
          },
          "translation_applicable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": false,
            "title": "Translation Applicable"
          }
        },
        "required": [
          "lov_name"
        ],
        "title": "LovCreate",
        "type": "object"
      },
      "LovUpdate": {
        "properties": {
          "apex_item_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apex Item Height"
          },
          "apex_item_width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apex Item Width"
          },
          "display_extra": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Extra"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "is_enterable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Enterable"
          },
          "is_static": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Static"
          },
          "is_translatable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Translatable"
          },
          "lov_definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lov Definition"
          },
          "lov_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lov Name"
          },
          "null_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Text"
          },
          "null_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Null Value"
          },
          "show_null_value": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Null Value"
          },
          "static_values": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Static Values"
          },
          "translation_applicable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Applicable"
          }
        },
        "title": "LovUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "API for building APEX-like applications",
    "title": "Open Source APEX Equivalent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/lovs/": {
      "get": {
        "description": "Retrieve LOVs.\nADMIN/DEVELOPER: Can access all LOVs\nEND_USER: Can only access LOVs in their current session's application",
        "operationId": "read_lovs_api_v1_lovs__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Lov"
                  },
                  "title": "Response Read Lovs Api V1 Lovs  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Lovs",
        "tags": [
          "lovs",
          "lovs"
        ]
      },
      "post": {
        "description": "Create new LOV.\nADMIN/DEVELOPER: Can create LOVs in any application\nEND_USER: Can only create LOVs in their current session's application",
        "operationId": "create_lov_api_v1_lovs__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LovCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lov"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Lov",
        "tags": [
          "lovs",
          "lovs"
        ]
      }
    },
    "/api/v1/lovs/name/{lov_name}": {
      "get": {
        "description": "Get LOV by name.\nADMIN/DEVELOPER: Can access any LOV\nEND_USER: Can only access LOVs from their current session's application",
        "operationId": "read_lov_by_name_api_v1_lovs_name__lov_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "lov_name",
            "required": true,
            "schema": {
              "title": "Lov Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lov"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Lov By Name",
        "tags": [
          "lovs",
          "lovs"
        ]
      }
    },
    "/api/v1/lovs/{lov_id}": {
      "delete": {
        "description": "Delete a LOV.\nOnly ADMIN role can delete LOVs.",
        "operationId": "delete_lov_api_v1_lovs__lov_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lov_id",
            "required": true,
            "schema": {
              "title": "Lov Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lov"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Lov",
        "tags": [
          "lovs",
          "lovs"
        ]
      },
      "get": {
        "description": "Get LOV by ID.\nADMIN/DEVELOPER: Can access any LOV\nEND_USER: Can only access LOVs from their current session's application",
        "operationId": "read_lov_api_v1_lovs__lov_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lov_id",
            "required": true,
            "schema": {
              "title": "Lov Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lov"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Lov",
        "tags": [
          "lovs",
          "lovs"
        ]
      },
      "put": {
        "description": "Update an LOV.\nADMIN/DEVELOPER: Can update any LOV\nEND_USER: Can only update LOVs in their current session's application",
        "operationId": "update_lov_api_v1_lovs__lov_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "lov_id",
            "required": true,
            "schema": {
              "title": "Lov Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LovUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lov"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Lov",
        "tags": [
          "lovs",
          "lovs"
        ]
      }
    },
...
  }
}